Move `ParamTerm` to `term.rs`
